requirements.txt
asgiref==3.2.7
backcall==0.1.0
colorama==0.4.3
ctypes-snappy==1.2
cypari==2.3.1
Cython==0.29.19
decorator==4.4.2
Django==3.0.7
djangorestframework==3.11.1
future==0.18.2
FXrays==1.3.3
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
networkx==2.4
numpy==1.18.2
pandas==1.0.3
parso==0.7.0
pickleshare==0.7.5
plink==2.3
prompt-toolkit==3.0.5
psycopg2==2.8.6
py4j==0.10.7
Pygments==2.6.1
pypandoc==1.5
pypng==0.0.20
pyreadline==2.1
pyspark==2.4.5
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
snappy==2.7
snappy-manifolds==1.1
spherogram==1.8.2
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.9
